+++ app/src/main/java/com/sitexa/android/community/presenter/UserDetailsPresenter.java
@@ -46,7 +46,7 @@ public class UserDetailsPresenter implements Presenter {
     /**
      * id used to retrieve user details
      */
-    private int userId;
+    private long userId;
     private UserDetailsView viewDetailsView;
 
     @Inject
@@ -76,7 +76,7 @@ public class UserDetailsPresenter implements Presenter {
     /**
      * Initializes the presenter by start retrieving user details.
      */
-    public void initialize(int userId) {
+    public void initialize(long userId) {
         this.userId = userId;
         this.loadUserDetails();
     }
