+++ data/src/main/java/com/sitexa/android/data/repository/datasource/UserDataStore.java
@@ -35,5 +35,5 @@ public interface UserDataStore {
    *
    * @param userId The id to retrieve user data.
    */
-  Observable<UserEntity> userEntityDetails(final int userId);
+  Observable<UserEntity> userEntityDetails(final long userId);
 }
